rename everything splitVcf into split_vcf
