+++ setup.py
@@ -53,7 +53,7 @@ setup(
         "somaticseq/utilities/dockered_pipelines/makeAlignmentScripts.py",
         "somaticseq/utilities/dockered_pipelines/makeSomaticScripts.py",
         "somaticseq/utilities/dockered_pipelines/run_workflows.py",
-        "somaticseq/vcf_modifier/splitVcf.py",
+        "somaticseq/vcf_modifier/split_vcf.py",
         "r_scripts/ada_model_builder_ntChange.R",
         "r_scripts/ada_model_predictor.R",
     ],
